requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
